+++ .eslint-bin/pre-commit-hook.js
@@ -3,6 +3,12 @@ const simpleGit = require('simple-git')(`${__dirname}/..`);
 
 let listFilesModified = [];
 
+function excludeNonCommitedFiles(file) {
+  return file.index !== 'D' // NOTICE: Deleted files
+    && file.index !== ' ' // NOTICE: Files not staged for commit
+    && file.index !== '?'; // NOTICE: Untracked files
+}
+
 function getFilesModified(callback) {
   simpleGit.status((error, status) => {
     if (error) {
@@ -12,7 +18,7 @@ function getFilesModified(callback) {
     }
 
     listFilesModified = status.files
-      .filter((change) => change.index !== 'D')
+      .filter(excludeNonCommitedFiles)
       .map((file) => file.path)
       .filter((file) => file.endsWith('.js'));