+++ test/fixtures/mysql/addresses.sql
@@ -1,7 +1,8 @@
-create table addresses (
-  id varchar(36) not null,
-  customer_id int unsigned null,
-  city varchar(127) not null,
-  constraint addresses_pk primary key (id),
-  constraint addresses_customers_id_fk foreign key (customer_id) references customers (id)
+CREATE TABLE addresses (
+  user_id INT NOT NULL PRIMARY KEY,
+  street VARCHAR(30) NOT NULL,
+  city VARCHAR(30) NOT NULL,
+  state VARCHAR(30) NOT NULL,
+  CONSTRAINT fk_user_id FOREIGN KEY (user_id) REFERENCES users (id),
+  CONSTRAINT unique_city UNIQUE (city)
 );