+++ test/fixtures/mysql/customers.sql
@@ -1,9 +1,9 @@
 create table customers (
-  id int unsigned auto_increment primary key,
-  name varchar(255) not null,
-  description text null,
-  is_active tinyint(1) default 1 not null,
-  paying bit(1) default null,
-  created_at date not null,
-  updated_at date null
+  id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+  name VARCHAR(255) NOT NULL,
+  description TEXT,
+  is_active BIT(1) DEFAULT 1 NOT NULL,
+  paying BIT(1) DEFAULT 0 NOT NULL,
+  created_at DATE NOT NULL,
+  updated_at DATE NOT NULL
 );