+++ test/fixtures/postgres/customers.sql
@@ -1,9 +1,9 @@
 create table customers (
-  id serial constraint customers_pk primary key,
-  name varchar not null,
-  description text,
-  is_active boolean default true not null,
-  paying bit(1) default null,
-  created_at timestamp not null,
-  updated_at timestamp
+  id SERIAL CONSTRAINT customers_pk PRIMARY KEY,
+  name VARCHAR NOT NULL,
+  description TEXT,
+  is_active BOOLEAN DEFAULT true NOT NULL,
+  paying BOOLEAN DEFAULT false NOT NULL,
+  created_at TIMESTAMP NOT NULL,
+  updated_at TIMESTAMP NOT NULL
 );