+++ test/fixtures/postgres/users.sql
@@ -1,0 +1,7 @@
+CREATE TABLE users (
+  id SERIAL,
+  username VARCHAR(25) NOT NULL,
+  enabled BOOLEAN DEFAULT TRUE,
+  last_login TIMESTAMP NOT NULL,
+  PRIMARY KEY (id)
+);